feat: error refinements, use reportError when available
