--- src/import-map.js
@@ -1,1 +1,0 @@
-export let importMap = { imports: {}, scopes: {} };
